docteurmemo/requirements.txt
django==4.2.7
python-dotenv==1.0.0
psycopg2-binary==2.9.7
djangorestframework==3.14.0
django-cors-headers==4.2.0
djangorestframework-simplejwt==5.3.0
requests==2.31.0
gunicorn==21.2.0
./docteurmemo-db-0.1.tar.gz